Restrict maximum icon size with --max-icon-size
By using the --max-icon-size option any icons that are larger than the
requested icon size get discarded. This is useful for at least Fvwm,
where the menus can hold icons of any size, with no option to resize to
a maximum size, making for some funny looking menus.

This adds a new dependency on the Pillow library, which is only loaded
if the --max-icon-size option is used.